update jest config
