--- jest.config.js
@@ -1,21 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const nextJest = require('next/jest');
-
-const createJestConfig = nextJest({
-    // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
-    dir: './'
-});
-
-// Add any custom config to be passed to Jest
-const customJestConfig = {
-    setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
-    moduleNameMapper: {
-        // Handle module aliases (this will be automatically configured for you soon)
-        '^@/(.*)$': '<rootDir>/src/$1'
-    },
-    testEnvironment: 'jest-environment-jsdom',
-    modulePathIgnorePatterns: ['<rootDir>/.next', '<rootDir>/tests']
-};
-
-// createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
-module.exports = createJestConfig(customJestConfig);
