+++ jest.setup.ts
@@ -1,8 +1,3 @@
-// Optional: configure or set up a testing framework before each test.
-// If you delete this file, remove `setupFilesAfterEnv` from `jest.config.js`
-
-// Used for __tests__/testing-library.js
-// Learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
 
 import jestFetchMock from 'jest-fetch-mock';
